requirements.txt

google-auth
google-auth-oauthlib
google-api-python-client
msal
requests
beautifulsoup4
